Modify link tooltip placement
Set the link tooltip placement to be 'auto' in general to prevent
tooltip from being cutoff when link is near the top of the textbox
in question player mode.

Also set the placement to be 'bottom' in exploration editor mode
in the Rich Text Editor (RTE) to prevent tooltip from covering
RTE components.